Move FileNote and Note interface #611
