+++ visualization/app/codeCharta/state/store/fileSettings/fileNotes/fileNotes.splitter.ts
@@ -1,5 +1,5 @@
 import { FileNotesAction, setFileNotes } from "./fileNotes.actions"
-import { FileNote } from "../../../../ui/attributeSideBar/attributeSideBar.component"
+import { FileNote } from "./fileNotes.service"
 
 export function splitFileNotesAction(payload: FileNote[]): FileNotesAction {
 	return setFileNotes(payload)
